B9PartSwitch compability
Well, I don't like Interstellar Fuel Switch very much because it's a little annoying to see a bunch of switchable resources I'd never use, so I made a B9partswitch config for myself to use a while ago. I checked that price/volume/mass are still the same for the ore tanks with Interstellar Fuel Switch installed, and the game works fine when both IFS and B9switch are installed.